requirements.txt
django
djangorestframework
psycopg2-binary
pillow
djangorestframework-simplejwt
python-decouple
django-cors-headers